Moved the model information to it's own file
