+++ internal/renderer/opengl_renderer.go
@@ -2,11 +2,9 @@ package renderer
 
 import (
 	"Gopher3D/internal/logger"
-	"bytes"
 	"fmt"
 	"image"
 	"image/draw"
-	"math"
 	"os"
 	"strings"
 
@@ -16,17 +14,19 @@ import (
 )
 
 type OpenGLRenderer struct {
-	Shader            Shader
-	Models            []*Model
-	vertexShader      uint32
-	fragmentShader    uint32
-	shaderProgram     uint32
-	modelLoc          int32
-	viewProjLoc       int32
-	lightPosLoc       int32
-	lightColorLoc     int32
-	lightIntensityLoc int32
-	Debug             bool
+	FrustumCullingEnabled bool
+	FaceCullingEnabled    bool
+	Debug                 bool
+	modelLoc              int32
+	viewProjLoc           int32
+	lightPosLoc           int32
+	lightColorLoc         int32
+	lightIntensityLoc     int32
+	vertexShader          uint32
+	fragmentShader        uint32
+	shaderProgram         uint32
+	Shader                Shader
+	Models                []*Model
 }
 
 func (rend *OpenGLRenderer) Init(width, height int32) {
@@ -34,7 +34,9 @@ func (rend *OpenGLRenderer) Init(width, height int32) {
 		logger.Log.Error("OpenGL initialization failed", zap.Error(err))
 		return
 	}
-	SetDefaultTexture()
+	rend.FrustumCullingEnabled = false
+	rend.FaceCullingEnabled = false
+	SetDefaultTexture(rend)
 	gl.Viewport(0, 0, width, height)
 	rend.Shader = InitShader()
 	rend.vertexShader = genShader(rend.Shader.vertexSource, gl.VERTEX_SHADER)
@@ -111,19 +113,28 @@ func (rend *OpenGLRenderer) Render(camera Camera, deltaTime float64, light *Ligh
 
 	// Get the uniform location of the texture sampler in your shader program
 	textureUniform := gl.GetUniformLocation(rend.shaderProgram, gl.Str("uTexture\x00"))
+
 	gl.Enable(gl.DEPTH_TEST)
+
 	// Culling : https://learnopengl.com/Advanced-OpenGL/Face-culling
-	gl.Enable(gl.CULL_FACE)
-	// IF FACES OF THE MODEL ARE RENDERED IN THE WRONG ORDER, TRY SWITCHING THE FOLLOWING LINE TO gl.CCW or we need to make sure the winding of each model is consistent
-	// CCW = Counter ClockWise
-	gl.CullFace(gl.FRONT)
-	gl.FrontFace(gl.CW)
+	if rend.FaceCullingEnabled {
+		gl.Enable(gl.CULL_FACE)
+		// IF FACES OF THE MODEL ARE RENDERED IN THE WRONG ORDER, TRY SWITCHING THE FOLLOWING LINE TO gl.CCW or we need to make sure the winding of each model is consistent
+		// CCW = Counter ClockWise
+		gl.CullFace(gl.FRONT)
+		gl.FrontFace(gl.CW)
+	}
 
 	// Calculate frustum
-	frustum := camera.CalculateFrustum()
+	var frustum Frustum
+
+	// TODO: Add check to see if camera is dirty(moved)
+	if rend.FrustumCullingEnabled {
+		frustum = camera.CalculateFrustum()
+	}
 	for _, model := range rend.Models {
 		// Skip rendering if the model is outside the frustum
-		if !frustum.IntersectsSphere(model.BoundingSphereCenter, model.BoundingSphereRadius) {
+		if rend.FrustumCullingEnabled && !frustum.IntersectsSphere(model.BoundingSphereCenter, model.BoundingSphereRadius) {
 			continue
 		}
 
@@ -174,6 +185,67 @@ func (rend *OpenGLRenderer) Cleanup() {
 		gl.DeleteBuffers(1, &model.VBO)
 		gl.DeleteBuffers(1, &model.EBO)
 	}
+}
+
+func (rend *OpenGLRenderer) LoadTexture(filePath string) (uint32, error) { // TODO: Consider specifying image format or handling different formats properly
+
+	imgFile, err := os.Open(filePath)
+	if err != nil {
+		return 0, err
+	}
+	defer imgFile.Close()
+
+	img, _, err := image.Decode(imgFile)
+	if err != nil {
+		return 0, err
+	}
+
+	rgba := image.NewRGBA(img.Bounds())
+	if rgba.Stride != rgba.Rect.Size().X*4 {
+		return 0, fmt.Errorf("unsupported stride")
+	}
+	draw.Draw(rgba, rgba.Bounds(), img, image.Point{0, 0}, draw.Src)
+
+	var textureID uint32
+	// Like any of the previous objects in OpenGL, textures are referenced with an ID; let's create one:
+	gl.GenTextures(1, &textureID)
+	gl.BindTexture(gl.TEXTURE_2D, textureID)
+	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(rgba.Rect.Size().X), int32(rgba.Rect.Size().Y), 0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(rgba.Pix))
+
+	// Set texture parameters (optional)
+	// gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_R, gl.REPEAT)
+
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+
+	// GL_NEAREST results in blocked patterns where we can clearly see the pixels that form the texture while GL_LINEAR produces a smoother pattern where the individual pixels are less visible.
+	// GL_LINEAR produces a more realistic output, but some developers prefer a more 8-bit look and as a result pick the GL_NEAREST option
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
+
+	return textureID, nil
+}
+
+func (rend *OpenGLRenderer) CreateTextureFromImage(img image.Image) (uint32, error) {
+	var textureID uint32
+	gl.GenTextures(1, &textureID)
+	gl.BindTexture(gl.TEXTURE_2D, textureID)
+
+	rgba, ok := img.(*image.RGBA)
+	if !ok {
+		// Convert to *image.RGBA if necessary
+		b := img.Bounds()
+		rgba = image.NewRGBA(image.Rect(0, 0, b.Dx(), b.Dy()))
+		draw.Draw(rgba, rgba.Bounds(), img, b.Min, draw.Src)
+	}
+
+	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(rgba.Rect.Size().X), int32(rgba.Rect.Size().Y), 0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(rgba.Pix))
+
+	// Set texture parameters
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
+
+	return textureID, nil
 }
 
 func genShader(source string, shaderType uint32) uint32 {
@@ -241,165 +313,4 @@ func CreateLight() *Light {
 		Color:     mgl32.Vec3{1.0, 1.0, 1.0},   // White light
 		Intensity: 1.0,                         // Full intensity
 	}
-}
-
-// TODO: This could be moved to a separate model package with a model interface
-func (m *Model) RotateModel(angleX, angleY float32, angleZ float32) {
-	// Create quaternions for each axis
-	rotationX := mgl32.QuatRotate(mgl32.DegToRad(angleX), mgl32.Vec3{1, 0, 0})
-	rotationY := mgl32.QuatRotate(mgl32.DegToRad(angleY), mgl32.Vec3{0, 1, 0})
-	rotationZ := mgl32.QuatRotate(mgl32.DegToRad(angleZ), mgl32.Vec3{0, 0, 1})
-
-	// Combine new rotation with existing rotation
-	m.Rotation = m.Rotation.Mul(rotationX).Mul(rotationY).Mul(rotationZ)
-	m.IsDirty = true
-}
-
-// SetPosition sets the position of the model
-func (m *Model) SetPosition(x, y, z float32) {
-	m.ModelMatrix = mgl32.Translate3D(x, y, z)
-	m.Position = mgl32.Vec3{x, y, z}
-	m.CalculateBoundingSphere()
-	m.IsDirty = true
-}
-
-func (m *Model) CalculateBoundingSphere() {
-	var center mgl32.Vec3
-	var maxDistanceSq float32
-
-	numVertices := len(m.Vertices) / 3 // Assuming 3 float32s per vertex
-	for i := 0; i < numVertices; i++ {
-		// Extracting vertex from the flat array
-		vertex := mgl32.Vec3{m.Vertices[i*3], m.Vertices[i*3+1], m.Vertices[i*3+2]}
-		transformedVertex := ApplyModelTransformation(vertex, m.Position, m.Scale, m.Rotation)
-		center = center.Add(transformedVertex)
-	}
-	center = center.Mul(1.0 / float32(numVertices))
-
-	for i := 0; i < numVertices; i++ {
-		vertex := mgl32.Vec3{m.Vertices[i*3], m.Vertices[i*3+1], m.Vertices[i*3+2]}
-		transformedVertex := ApplyModelTransformation(vertex, m.Position, m.Scale, m.Rotation)
-		distanceSq := transformedVertex.Sub(center).LenSqr()
-		if distanceSq > maxDistanceSq {
-			maxDistanceSq = distanceSq
-		}
-	}
-
-	m.BoundingSphereCenter = center
-	m.BoundingSphereRadius = float32(math.Sqrt(float64(maxDistanceSq)))
-}
-
-func (m *Model) SetTexture(texturePath string) {
-	textureID, err := loadTexture(texturePath)
-	if err != nil {
-		logger.Log.Error("Failed to load texture", zap.String("path", texturePath), zap.Error(err))
-		return
-	}
-
-	if m.Material == nil {
-		logger.Log.Info("Setting default material")
-		m.Material = DefaultMaterial
-
-	}
-	m.Material.TextureID = textureID
-}
-
-// Aux functions, maybe I need to move them to another package
-func ApplyModelTransformation(vertex, position, scale mgl32.Vec3, rotation mgl32.Quat) mgl32.Vec3 {
-	// Apply scaling
-	scaledVertex := mgl32.Vec3{vertex[0] * scale[0], vertex[1] * scale[1], vertex[2] * scale[2]}
-
-	// Apply rotation
-	// Note: mgl32.Quat doesn't directly multiply with Vec3, so we convert it to a Mat4 first
-	rotatedVertex := rotation.Mat4().Mul4x1(scaledVertex.Vec4(1)).Vec3()
-
-	// Apply translation
-	transformedVertex := rotatedVertex.Add(position)
-
-	return transformedVertex
-}
-
-func SetDefaultTexture() {
-	// Read the embedded texture
-	textureBytes, err := defaultTextureFS.ReadFile("resources/default.png")
-	if err != nil {
-		logger.Log.Error("Failed to read embedded default texture", zap.Error(err))
-		return
-	}
-
-	// Create an image from the texture bytes
-	img, _, err := image.Decode(bytes.NewReader(textureBytes))
-	if err != nil {
-		logger.Log.Error("Failed to decode embedded default texture", zap.Error(err))
-		return
-	}
-
-	// Convert the image to a texture and set it as the default texture
-	textureID, err := createTextureFromImage(img) // Assumes this function exists
-	if err != nil {
-		logger.Log.Error("Failed to create texture from embedded default image", zap.Error(err))
-		return
-	}
-
-	DefaultMaterial.TextureID = textureID
-}
-func createTextureFromImage(img image.Image) (uint32, error) {
-	var textureID uint32
-	gl.GenTextures(1, &textureID)
-	gl.BindTexture(gl.TEXTURE_2D, textureID)
-
-	rgba, ok := img.(*image.RGBA)
-	if !ok {
-		// Convert to *image.RGBA if necessary
-		b := img.Bounds()
-		rgba = image.NewRGBA(image.Rect(0, 0, b.Dx(), b.Dy()))
-		draw.Draw(rgba, rgba.Bounds(), img, b.Min, draw.Src)
-	}
-
-	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(rgba.Rect.Size().X), int32(rgba.Rect.Size().Y), 0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(rgba.Pix))
-
-	// Set texture parameters
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
-
-	return textureID, nil
-}
-
-func loadTexture(filePath string) (uint32, error) { // TODO: Consider specifying image format or handling different formats properly
-
-	imgFile, err := os.Open(filePath)
-	if err != nil {
-		return 0, err
-	}
-	defer imgFile.Close()
-
-	img, _, err := image.Decode(imgFile)
-	if err != nil {
-		return 0, err
-	}
-
-	rgba := image.NewRGBA(img.Bounds())
-	if rgba.Stride != rgba.Rect.Size().X*4 {
-		return 0, fmt.Errorf("unsupported stride")
-	}
-	draw.Draw(rgba, rgba.Bounds(), img, image.Point{0, 0}, draw.Src)
-
-	var textureID uint32
-	// Like any of the previous objects in OpenGL, textures are referenced with an ID; let's create one:
-	gl.GenTextures(1, &textureID)
-	gl.BindTexture(gl.TEXTURE_2D, textureID)
-	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(rgba.Rect.Size().X), int32(rgba.Rect.Size().Y), 0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(rgba.Pix))
-
-	// Set texture parameters (optional)
-	// gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_R, gl.REPEAT)
-
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
-
-	// GL_NEAREST results in blocked patterns where we can clearly see the pixels that form the texture while GL_LINEAR produces a smoother pattern where the individual pixels are less visible.
-	// GL_LINEAR produces a more realistic output, but some developers prefer a more 8-bit look and as a result pick the GL_NEAREST option
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
-	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
-
-	return textureID, nil
 }
